write down parser

+++ src/main/kotlin/com/cepgamer/strattester/security/PriceCandle.kt
@@ -9,6 +9,6 @@ data class PriceCandle(
     val low: BigDecimal,
     val high: BigDecimal,
     val volume: BigDecimal,
-    val openTimestamp: BigInteger,
-    val timespan: BigInteger
+    val openTimestamp: Int,
+    val timespan: Int
 )
